fix: move GitHubRelease type to settingsInterfaces and update OpenAPI docs
