+++ server/api/github.ts
@@ -1,27 +1,9 @@
 import cacheManager from '@server/lib/cache';
+import type { GitHubRelease } from '@server/interfaces/api/settingsInterfaces';
 import logger from '@server/logger';
 import ExternalAPI from './externalapi';
 
 const GITHUB_CACHE_TTL = 1800;
-
-export interface GitHubRelease {
-  url: string;
-  assets_url: string;
-  upload_url: string;
-  html_url: string;
-  id: number;
-  node_id: string;
-  tag_name: string;
-  target_commitish: string;
-  name: string;
-  draft: boolean;
-  prerelease: boolean;
-  created_at: string;
-  published_at: string;
-  tarball_url: string;
-  zipball_url: string;
-  body: string;
-}
 
 interface GithubCommit {
   sha: string;
