+++ server/interfaces/api/settingsInterfaces.ts
@@ -87,3 +87,22 @@ export interface PythonServiceStatusResponse {
   lastHealthy: string | null;
   consecutiveFailures: number;
 }
+
+export interface GitHubRelease {
+  url: string;
+  assets_url: string;
+  upload_url: string;
+  html_url: string;
+  id: number;
+  node_id: string;
+  tag_name: string;
+  target_commitish: string;
+  name: string;
+  draft: boolean;
+  prerelease: boolean;
+  created_at: string;
+  published_at: string;
+  tarball_url: string;
+  zipball_url: string;
+  body: string;
+}
