+++ streamarr-api.yml
@@ -3510,9 +3510,12 @@ paths:
   /settings/releases:
     get:
       summary: Get GitHub releases
-      description: Returns a list of Streamarr releases from GitHub, served through the server-side cache.
+      description: Returns a list of Streamarr releases from GitHub, served through the server-side cache. Requires admin authentication.
       tags:
         - settings
+      security:
+        - cookieAuth: []
+        - apiKey: []
       responses:
         '200':
           description: List of releases
@@ -3525,11 +3528,19 @@ paths:
                   properties:
                     url:
                       type: string
+                    assets_url:
+                      type: string
+                    upload_url:
+                      type: string
                     html_url:
                       type: string
                     id:
                       type: number
+                    node_id:
+                      type: string
                     tag_name:
+                      type: string
+                    target_commitish:
                       type: string
                     name:
                       type: string
@@ -3541,8 +3552,17 @@ paths:
                       type: string
                     published_at:
                       type: string
+                    tarball_url:
+                      type: string
+                    zipball_url:
+                      type: string
                     body:
                       type: string
+                  additionalProperties: true
+        '401':
+          description: Unauthorized - user not authenticated
+        '403':
+          description: Forbidden - user lacks admin permission
   /settings/restart-required:
     get:
       summary: Check if server restart is required
